requirements.txt
python-dotenv
requests
google-generativeai
twilio
supabase
anthropic
openai
